root/requirements_nv.txt
# mysqlclient==1.4.6
channels==2.4.0
cryptography==2.8
Django==2.2.24
django-cors-headers==3.2.1
django-haystack==3.2.1
django-health-check==3.14.0
django-rest-passwordreset==1.1.0
django-ses==1.0.3
django-storages==1.9.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
fcm-django==0.3.4
gensim==3.8.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
oauth2client==4.1.3
Pillow==6.2.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
Whoosh==2.7.4
watchtower==0.7.3
boto3==1.12.39
aioredis==1.3.1
asgiref==3.2.3
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
boto==2.49.0
certifi==2019.11.28
cffi==1.13.2
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
daphne==2.5.0
django-redis==4.11.0
django-templated-mail==1.1.1
djoser==2.0.3
future==0.18.2
fuzzywuzzy==0.18.0
gunicorn==20.0.4
hiredis==1.0.1
httpie==1.0.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
pycparser==2.19
Pygments==2.5.2
PyHamcrest==1.9.0
python-Levenshtein==0.12.0
# python-slugify
# pywin32
redis==3.4.1
requests==2.22.0
slugify==0.0.1
text-unidecode==1.3
Twisted==19.10.0
txaio==18.8.1
# urllib3
wincertstore==0.2
# zope.interface
# django-extensions
mysqlclient==2.1.1
backports.zoneinfo==0.2.1
opensearch-py==2.5.0